linux-requirements.txt
numpy
matplotlib
pandas
tqdm
certifi
pyyaml
scipy
cleverhans

# if on linux, run:
# conda install -c pytorch torchvision 
# check to see if cudatoolkit was installed with 'conda list'.  
# if it was not installed, run:
# conda install cudatoolkit
# install the requirements in this file with
# conda install --file linux-requirements.txt

# to "install" shrinkbench, go to ~/git and run
# git clone https://github.com/JJGO/shrinkbench.git
# conda activate <environment_name>
# conda env config vars set PYTHONPATH="<path to parent folder of shrinkbench>"
# conda activate <environment_name>
# you can check that this worked with
# echo $PYTHONPATH
